+++ examples/info.rs
@@ -6,7 +6,8 @@ fn main() {
     let manager = Manager::new();
 
     for loader in manager.get_loader_list().iter() {
-        let extensions: String = loader.get_extensions().join(",");
+        let extensions: Vec<String> = loader.get_extensions_mime_type_map().keys().map(|s| s.clone()).collect();
+        let extensions: String = extensions.join(",");
         let mime_types: String = loader.get_mime_types().join(",");
         println!(
             "Loader {}: Extensions=[{}], Mime-Types=[{}] ",