+++ examples/to_x3d.rs
@@ -1,23 +1,27 @@
-use std::{collections::HashSet, env, path::Path};
+use std::{env, fs::File, path::Path};
 
 use cad_import::{
+    exporter::X3DExporter,
     loader::{FileResource, Manager},
-    structure::{CADData, Node},
-    ID,
 };
 
-///! This example loads the given file and dumps all infos about it to console
+///! This example loads the given file and exports as X3D to the specified path
 
 fn usage() {
-    println!("usage: file_info <file-path> [<mime-type>]\n");
+    println!("usage: to_x3d <file-path> <x3d-path> [<mime-type>]\n");
     println!("file-path: The path to the cad file to parse.");
+    println!("x3d-path: The path to the resulting X3D file.");
     println!(
         "mime-type: Optional parameter to specify the mimetype. If not provided, the extension
            of the file is used"
     );
 }
 
-fn determine_mime_types(manager: &Manager, input_file: &Path, mime_type: Option<&str>) -> Vec<String> {
+fn determine_mime_types(
+    manager: &Manager,
+    input_file: &Path,
+    mime_type: Option<&str>,
+) -> Vec<String> {
     match mime_type {
         Some(m) => vec![m.to_owned()],
         None => match input_file.extension() {
@@ -36,53 +40,7 @@ fn determine_mime_types(manager: &Manager, input_file: &Path, mime_type: Option<
     }
 }
 
-struct VisitorContext {
-    pub num_nodes: usize,
-    pub num_vertices: usize,
-    pub num_primitives: usize,
-    pub shapes: HashSet<ID>,
-}
-
-fn visit_node(node: &Node, ctx: &mut VisitorContext) {
-    ctx.num_nodes += 1;
-
-    for shape in node.get_shapes().iter() {
-        let id = shape.get_id();
-        if ctx.shapes.contains(&id) {
-            continue;
-        }
-
-        ctx.shapes.insert(id);
-
-        for part in shape.get_parts() {
-            let mesh = part.get_mesh();
-            ctx.num_vertices += mesh.get_vertices().len();
-            ctx.num_primitives += mesh.get_primitives().num_primitives();
-        }
-    }
-
-    for child in node.get_children().iter() {
-        visit_node(child, ctx);
-    }
-}
-
-fn dump_info(cad_data: &CADData) {
-    let mut ctx = VisitorContext {
-        num_nodes: 0,
-        num_vertices: 0,
-        num_primitives: 0,
-        shapes: HashSet::new(),
-    };
-    visit_node(cad_data.get_root_node(), &mut ctx);
-
-    println!("Statistics:");
-    println!("Num Vertices: {}", ctx.num_vertices);
-    println!("Num Primitives: {}", ctx.num_primitives);
-    println!("Num Nodes: {}", ctx.num_nodes);
-    println!("Num Shapes: {}", ctx.shapes.len());
-}
-
-fn run_program(input_file: &Path, mime_type: Option<&str>) -> bool {
+fn run_program(input_file: &Path, x3d_file: &Path, mime_type: Option<&str>) -> bool {
     let manager = Manager::new();
 
     let mime_types = determine_mime_types(&manager, input_file, mime_type);
@@ -114,7 +72,27 @@ fn run_program(input_file: &Path, mime_type: Option<&str>) -> bool {
 
     println!("Reading file {:?}...DONE", input_file);
 
-    dump_info(&cad_data);
+    println!("Writing X3D {:?}...", x3d_file);
+    let x3d_exporter = X3DExporter::new(&cad_data);
+    let file = match File::create(x3d_file) {
+        Ok(f) => f,
+        Err(err) => {
+            eprintln!("Writing X3D {:?}...FAILED", x3d_file);
+            eprintln!("Error: {}", err);
+            return false;
+        }
+    };
+
+    match x3d_exporter.write(file) {
+        Ok(()) => {
+            println!("Writing X3D {:?}...DONE", x3d_file);
+        }
+        Err(err) => {
+            eprintln!("Writing X3D {:?}...FAILED", x3d_file);
+            eprintln!("Error: {}", err);
+            return false;
+        }
+    }
 
     true
 }
@@ -124,10 +102,10 @@ fn main() {
     let args = &args[1..];
 
     // check if the number of arguments is invalid
-    if args.len() == 0 {
+    if args.len() < 2 {
         usage();
         std::process::exit(-1);
-    } else if args.len() > 2 {
+    } else if args.len() > 3 {
         eprintln!("Too many arguments!!!\n\n");
         usage();
         std::process::exit(-1);
@@ -135,13 +113,14 @@ fn main() {
 
     // parse arguments
     let input_file = Path::new(&args[0]);
-    let mime_type = if args.len() == 2 {
-        Some(args[1].as_str())
+    let output_file = Path::new(&args[1]);
+    let mime_type = if args.len() == 3 {
+        Some(args[2].as_str())
     } else {
         None
     };
 
-    if run_program(input_file, mime_type) {
+    if run_program(input_file, output_file, mime_type) {
         println!("FINISHED");
     } else {
         eprintln!("FAILED!!!");