+++ src/exporter/mod.rs
@@ -1,0 +1,5 @@
+//! The exporter module contains exporter functionalities to export the in-memory data.
+
+mod x3d_exporter;
+
+pub use x3d_exporter::X3DExporter;