+++ src/loader/loader.rs
@@ -1,4 +1,10 @@
+use std::{path::Path, collections::{BTreeMap, BTreeSet}};
+
 use crate::{error::Error, structure::CADData};
+
+use super::{FileResource, Resource};
+
+pub type ExtensionMap = BTreeMap<String, BTreeSet<String>>;
 
 /// The trait for a registered loader.
 pub trait Loader {
@@ -9,10 +15,10 @@ pub trait Loader {
     /// loader is chosen if multiple loaders match for a given resource.
     fn get_priority(&self) -> u32;
 
-    /// Returns list of supported file extensions in lower case.
-    fn get_extensions(&self) -> Vec<String>;
+    /// Returns a map of known file extensions with their respective mime types for the loader.
+    fn get_extensions_mime_type_map(&self) -> ExtensionMap;
 
-    /// Returns list of supported mime types for the format.
+    /// Returns list of supported mime types for the formats the loader is supporting.
     fn get_mime_types(&self) -> Vec<String>;
 
     /// Reads the CAD data from the given reader. If something happens, the loader will return
@@ -20,5 +26,15 @@ pub trait Loader {
     ///
     /// # Arguments
     /// * `reader` - The reader from which the loader will read the cad data.
-    fn read_file(&self, reader: &mut dyn std::io::Read) -> Result<CADData, Error>;
+    fn read(&self, resource: &dyn Resource) -> Result<CADData, Error>;
+
+    /// Reads the CAD data from the given path. If something happens, the loader will return
+    /// a error message.
+    ///
+    /// # Arguments
+    /// * `p` - The path from which the loader will read the cad data.
+    fn read_file(&self, p: &Path, mime_type: &str) -> Result<CADData, Error> {
+        let f = FileResource::new(p.to_owned(), mime_type);
+        self.read(&f)
+    }
 }