+++ src/loader/loader_gltf/mod.rs
@@ -1,0 +1,6 @@
+mod accessor_iterator;
+mod component;
+mod loader_gltf;
+mod utils;
+
+pub use loader_gltf::LoaderGLTF;