+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unpublished]
 
+## [0.3.0]
+
+### Added
+- New resource interface for accessing resources
+- Support for non-indexed primitives
+- Added export functionality to the X3D format
+- Added example program to convert data to X3D
+- Added GLTF loader
+
+### Changed
+- Changed the resource extension handling
+
 ## [0.2.1] - 2023-02-22
 
 ### Changed